Refactor TimeZoneHelper and introduce new utility classes for improved time zone handling and SQL string sanitization. Migrate time zone methods to the Utilities namespace, enhance XML documentation, and add DiagnosticConfig and SqlStringParam classes for better organization and security. Clean up using directives in BaseRepository for clarity.
