+++ src/Acontplus.Persistence.SqlServer/Utilities/DiagnosticConfig.cs
@@ -1,6 +1,6 @@
 ﻿using System.Diagnostics;
 
-namespace Acontplus.Persistence.SqlServer.Diagnostics;
+namespace Acontplus.Persistence.SqlServer.Utilities;
 
 public static class DiagnosticConfig
 {
